Create README.md
I believe data analysis is underrated vis a vis data science. I believe the best value comes from working at speed and in close alignment with business or organizational outcomes. I believe that art and technology are a good mix and that this is no exception in data. Making simple, easy-to-understand, data interpretations free of bias is the end product.